refactor: auth 도메인 페키지를 domain 페키지 하위로 이동
