+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/auth/service/RefreshTokenService.java
@@ -1,4 +1,4 @@
-package com.hertz.hertz_be.auth.service;
+package com.hertz.hertz_be.domain.auth.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.StringRedisTemplate;
