+++ hertz-be/src/main/java/com/hertz/hertz_be/global/exception/GlobalExceptionHandler.java
@@ -1,11 +1,5 @@
 package com.hertz.hertz_be.global.exception;
 
-import com.hertz.hertz_be.auth.exception.RefreshTokenInvalidException;
-import com.hertz.hertz_be.global.common.ResponseCode;
-import com.hertz.hertz_be.global.common.ResponseDto;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @RestControllerAdvice
